Agno_Groq/src/agent_runner.py
import os
from dotenv import load_dotenv

from agno.agent import Agent
from agno.models.groq import Groq
from agno.tools.duckduckgo import DuckDuckGoTools

load_dotenv()  # Loads from .env

agent = Agent(
    model=Groq(id="llama-3.3-70b-versatile", api_key=os.getenv("GROQ_API_KEY")),
    description="You are an Sports news reporter with a flair for Story Telling!",
    tools=[DuckDuckGoTools()],
    show_tool_calls=True,
    markdown=True,
)

agent.print_response("Tell me about a breaking news story from Latest Rajastan vs Gujarat IPL match", stream=True)
