Agno_Groq/pyproject.toml
[project]
name = "agno-project"
version = "0.1.0"
dependencies = [
    "agno>=1.4.2",
    "duckduckgo-search>=8.0.1",
    "groq>=0.23.1",
    "python-dotenv>=1.1.0",
]